Added example of encoding values
Added a Example that converts a number value into a string, and vice versa based on a basestring that you provide.
Can both encode and decode.
Can be used to take less space and to easier save data, but also to obfiscate the numbers to make it slightly harder to "cheat"